routes/api.py
from flask import Flask, Blueprint, render_template, request, redirect, jsonify, flash, session, url_for, Response
from datetime import date
from models import db, Event, Student, Attendance
import csv
from sqlalchemy import text
from models import db, Event, Student, Attendance
from functools import wraps
import time

api_bp = Blueprint('api', __name__)

